requirements.txt
polars>=0.20.0
pandas>=2.0.0
pyarrow>=14.0.0
snowflake-connector-python>=3.0.0
python-dotenv>=1.0.0
streamlit>=1.30.0
pyodbc>=5.0.0
